Handle Gauss() return value correctly
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/branches/S0.8@810 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ src/mathlib.cpp
@@ -61,7 +61,7 @@ TVector3d TransformPoint(const TMatrix<4, 4>& mat, const TVector3d& p) {
 bool IntersectPlanes(const TPlane& s1, const TPlane& s2, const TPlane& s3, TVector3d *p) {
 	double A[3][4];
 	double x[3];
-	double retval;
+	int retval;
 
 	A[0][0] =  s1.nml.x;
 	A[0][1] =  s1.nml.y;
